added estimate env temp for X19 and change format of X19 and X17 files

+++ data/__init__.py
@@ -40,7 +40,7 @@ class MinerData:
     hostname: str = "Unknown"
     hashrate: float = 0
     temperature_avg: int = field(init=False)
-    env_temp: int = 0
+    env_temp: float = 0
     left_board_temp: int = 0
     left_board_chip_temp: int = 0
     center_board_temp: int = 0
